Move getIndex again
