+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/Placeholders.java
@@ -144,44 +144,4 @@ public class Placeholders<T> {
         return placeholders;
     }
 
-    public static int getIndex(String text) {
-        StringBuilder sb = new StringBuilder();
-        boolean open = false;
-        boolean start = false;
-        int result = -1;
-
-        for (char c : text.toCharArray()) {
-            boolean end = false;
-
-            switch (c) {
-                case '{':
-                    open = true;
-                    break;
-                case '-':
-                    start = true;
-                    break;
-                case '}':
-                    end = true;
-                    break;
-                default:
-                    if (open && start) {
-                        sb.append(c);
-                    }
-            }
-
-            if (end) {
-                break;
-            }
-        }
-
-        try {
-            result = Integer.parseInt(sb.toString());
-        }
-        catch (NumberFormatException e) {
-            FunnyGuilds.getPluginLogger().parser(text + " contains an invalid number: " + sb.toString());
-        }
-
-        return result;
-    }
-
 }
